backend/requirements.txt
# Azure Functions dependencies
azure-functions>=1.20.0
azure-storage-blob>=12.14.0

# Data analysis dependencies
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
numpy>=1.21.0

# Redis cache
redis>=4.5.0